(#2889) implements a draft of table formatted output for choco outdated
The output of `choco outdated`is not easy to read for humans, the
readability gets enhanced by this commit, as it now is shown as table.

Only noticeable difference is, that the output now only will be visible
after all outdated packages were obtained, as we need to know the
length of the texts before we could output the list. Earlier the output
became visible package by package.